+++ bin/dhttpd.dart
@@ -1,0 +1,5 @@
+import '_boot.dart';
+
+main(List<String> args) async {
+  bootServer(args);
+}